Revert "Update to 2.0.0"
This reverts commit 8c039840d8d6729590ec8c7a3a673a1d124e0d3e.
