+++ exercises/clock/src/test/kotlin/ClockAddTest.kt
@@ -68,6 +68,68 @@ class ClockAddTest {
         assertEquals("11:21", clock.toString())
     }
 
+    @Ignore
+    @Test
+    fun subtractMinutes() {
+        val clock = Clock(10, 3)
+        clock.add(-3)
+        assertEquals("10:00", clock.toString())
+    }
 
+    @Ignore
+    @Test
+    fun subtractToPreviousHour() {
+        val clock = Clock(10, 3)
+        clock.add(-30)
+        assertEquals("09:33", clock.toString())
+    }
+
+    @Ignore
+    @Test
+    fun subtractMoreThanAnHour() {
+        val clock = Clock(10, 3)
+        clock.add(-70)
+        assertEquals("08:53", clock.toString())
+    }
+
+    @Ignore
+    @Test
+    fun subtractAcrossMidnight() {
+        val clock = Clock(0, 3)
+        clock.add(-4)
+        assertEquals("23:59", clock.toString())
+    }
+
+    @Ignore
+    @Test
+    fun subtractMoreThanTwoHours() {
+        val clock = Clock(0, 0)
+        clock.add(-160)
+        assertEquals("21:20", clock.toString())
+    }
+
+    @Ignore
+    @Test
+    fun subtractMoreThanTwoHoursWithBorrow() {
+        val clock = Clock(6, 15)
+        clock.add(-160)
+        assertEquals("03:35", clock.toString())
+    }
+
+    @Ignore
+    @Test
+    fun subtractMoreThanOneDay() {
+        val clock = Clock(5, 32)
+        clock.add(-1500)
+        assertEquals("04:32", clock.toString())
+    }
+
+    @Ignore
+    @Test
+    fun subtractMoreThanTwoDays() {
+        val clock = Clock(2, 20)
+        clock.add(-3000)
+        assertEquals("00:20", clock.toString())
+    }
 
 }
